+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/FrameworkExtension.php
@@ -965,6 +965,7 @@ class FrameworkExtension extends Extension
 
         // Use the "real" translator instead of the identity default
         $container->setAlias('translator', 'translator.default');
+        $container->setAlias('translator.formatter', new Alias($config['formatter'], false));
         $translator = $container->findDefinition('translator.default');
         $translator->addMethodCall('setFallbackLocales', array($config['fallbacks']));