+++ src/Symfony/Bundle/FrameworkBundle/Translation/Translator.php
@@ -14,8 +14,8 @@ namespace Symfony\Bundle\FrameworkBundle\Translation;
 use Psr\Container\ContainerInterface;
 use Symfony\Component\HttpKernel\CacheWarmer\WarmableInterface;
 use Symfony\Component\Translation\Translator as BaseTranslator;
-use Symfony\Component\Translation\MessageSelector;
 use Symfony\Component\Translation\Exception\InvalidArgumentException;
+use Symfony\Component\Translation\Formatter\MessageFormatterInterface;
 
 /**
  * Translator.
@@ -55,15 +55,15 @@ class Translator extends BaseTranslator implements WarmableInterface
      *   * debug:     Whether to enable debugging or not (false by default)
      *   * resource_files: List of translation resources available grouped by locale.
      *
-     * @param ContainerInterface $container     A ContainerInterface instance
-     * @param MessageSelector    $selector      The message selector for pluralization
-     * @param string             $defaultLocale
-     * @param array              $loaderIds     An array of loader Ids
-     * @param array              $options       An array of options
+     * @param ContainerInterface        $container     A ContainerInterface instance
+     * @param MessageFormatterInterface $formatter     The message formatter
+     * @param string                    $defaultLocale
+     * @param array                     $loaderIds     An array of loader Ids
+     * @param array                     $options       An array of options
      *
      * @throws InvalidArgumentException
      */
-    public function __construct(ContainerInterface $container, MessageSelector $selector, string $defaultLocale, array $loaderIds = array(), array $options = array())
+    public function __construct(ContainerInterface $container, $formatter, string $defaultLocale, array $loaderIds = array(), array $options = array())
     {
         $this->container = $container;
         $this->loaderIds = $loaderIds;
@@ -77,7 +77,7 @@ class Translator extends BaseTranslator implements WarmableInterface
         $this->resourceLocales = array_keys($this->options['resource_files']);
         $this->addResourceFiles($this->options['resource_files']);
 
-        parent::__construct($defaultLocale, $selector, $this->options['cache_dir'], $this->options['debug']);
+        parent::__construct($defaultLocale, $formatter, $this->options['cache_dir'], $this->options['debug']);
     }
 
     /**